re arranged files for notifications card
