+++ web/app/[workspaceSlug]/(projects)/notifications/layout.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 // components
-import { NotificationsSidebar } from "@/components/workspace-notifications";
+import { NotificationsSidebarRoot } from "@/components/workspace-notifications";
 
 export default function ProjectInboxIssuesLayout({ children }: { children: React.ReactNode }) {
   return (
     <div className="relative w-full h-full overflow-hidden flex items-center">
-      <NotificationsSidebar />
+      <NotificationsSidebarRoot />
       <div className="w-full h-full overflow-hidden overflow-y-auto">{children}</div>
     </div>
   );
