+++ web/ee/components/workspace-notifications/index.ts
@@ -1,0 +1,1 @@
+export * from './sidebar'
