+++ web/ee/components/workspace-notifications/sidebar/index.ts
