+++ web/ee/components/workspace-notifications/sidebar/notification-card/index.ts
@@ -1,0 +1,3 @@
+export * from "./root";
+export * from "./item";
+export * from "./options";
