+++ web/ee/components/workspace-notifications/sidebar/notification-card/item.tsx
@@ -1,0 +1,1 @@
+export  { NotificationItem } from "ce/components/workspace-notifications";
