+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/archive.tsx
@@ -1,0 +1,1 @@
+export { NotificationItemArchiveOption } from "ce/components/workspace-notifications";
