+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/button.tsx
@@ -1,0 +1,1 @@
+export { NotificationItemOptionButton } from "@/plane-web/components/workspace-notifications";
