+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/read.tsx
@@ -1,0 +1,1 @@
+export { NotificationItemReadOption } from "ce/components/workspace-notifications";
