+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/root.tsx
@@ -1,0 +1,1 @@
+export  { NotificationOption } from "ce/components/workspace-notifications";
