+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/snooze/modal.tsx
@@ -1,0 +1,1 @@
+export { NotificationSnoozeModal } from "ce/components/workspace-notifications";
