+++ web/ee/components/workspace-notifications/sidebar/notification-card/options/snooze/root.tsx
@@ -1,0 +1,1 @@
+export { NotificationItemSnoozeOption } from "ce/components/workspace-notifications";
