+++ web/ee/components/workspace-notifications/sidebar/notification-card/root.tsx
@@ -1,0 +1,1 @@
+export  { NotificationCardListRoot } from "ce/components/workspace-notifications";
