+++ web/ee/components/workspace-notifications/sidebar/root.tsx
@@ -1,0 +1,1 @@
+export { NotificationsSidebar } from "ce/components/workspace-notifications"
